+++ src/test/java/supernova/whokie/answer/AnswerIntegrationTest.java
@@ -20,7 +20,7 @@ import supernova.whokie.question.repository.QuestionRepository;
 import supernova.whokie.user.Gender;
 import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
-import supernova.whokie.user.repository.UserRepository;
+import supernova.whokie.user.infrastructure.repository.UserRepository;
 
 import java.time.LocalDateTime;
